+++ src/app/components/breadcrumbs/terra-breadcrumb-container.ts
@@ -1,0 +1,13 @@
+import { TerraBreadcrumb } from './terra-breadcrumb';
+
+export class TerraBreadcrumbContainer
+{
+    public currentSelectedBreadcrumb:TerraBreadcrumb;
+    public breadcrumbList:Array<TerraBreadcrumb>;
+    public isHidden:boolean;
+
+    constructor()
+    {
+        this.breadcrumbList = [];
+    }
+}