+++ src/app/components/breadcrumbs/terra-breadcrumbs.component.ts
@@ -1,0 +1,91 @@
+import { Component } from '@angular/core';
+import { TerraBreadcrumbsService } from './service/terra-breadcrumbs.service';
+import {
+    ActivatedRoute,
+    Route,
+    Router,
+    Routes
+} from '@angular/router';
+import { isNullOrUndefined } from 'util';
+import { TerraBreadcrumb } from './terra-breadcrumb';
+import { TerraBreadcrumbContainer } from './terra-breadcrumb-container';
+
+@Component(
+    {
+        selector:  'terra-breadcrumbs',
+        styles:    [require('./terra-breadcrumbs.component.scss')],
+        template:  require('./terra-breadcrumbs.component.html'),
+        providers: [TerraBreadcrumbsService]
+    }
+)
+export class TerraBreadcrumbsComponent
+{
+    constructor(private _breadcrumbsService:TerraBreadcrumbsService, private activatedRoute:ActivatedRoute, private router:Router)
+    {
+        this._breadcrumbsService.initialPath =
+            this.getCompletePathByRoute(this.activatedRoute.snapshot.routeConfig, this.router.config, '');
+    }
+
+    // identical code exists in TerraRouterHelper in terra
+    private getCompletePathByRoute(routeToFind:Route, routeConfig:Routes, parentPath:string):string
+    {
+        let path:string = null;
+
+        if(!isNullOrUndefined(routeConfig))
+        {
+            for(let route of routeConfig)
+            {
+                if(!isNullOrUndefined(route.data) && !isNullOrUndefined(route.data.permissions) && route === routeToFind)
+                {
+                    path = parentPath + '/' + route.path;
+
+                    if(parentPath.length === 0)
+                    {
+                        path = route.path;
+                    }
+
+                    return path;
+                }
+                else if(route.children)
+                {
+                    path = this.getCompletePathByRoute(routeToFind, route.children, parentPath);
+
+                    if(!isNullOrUndefined(path))
+                    {
+                        path = route.path + '/' + path;
+                        break;
+                    }
+                }
+            }
+        }
+
+        return path;
+    }
+
+    protected get breadcrumbContainer():Array<TerraBreadcrumbContainer>
+    {
+        return this._breadcrumbsService.breadcrumbContainer;
+    }
+
+    protected closeBreadcrumb(container:TerraBreadcrumbContainer, breadcrumb:TerraBreadcrumb, event:Event):void
+    {
+        event.stopPropagation();
+
+        this._breadcrumbsService.closeBreadcrumb(container, breadcrumb);
+    }
+
+    protected checkActiveRoute(bcc:TerraBreadcrumb):boolean
+    {
+        return this._breadcrumbsService.checkActiveRoute(bcc);
+    }
+
+    protected getAmountOfVisibleBreadcrumbsForContainer(container:TerraBreadcrumbContainer):number
+    {
+        return container.breadcrumbList.filter((bc:TerraBreadcrumb) => !bc.isHidden).length;
+    }
+
+    public get breadcrumbsService():TerraBreadcrumbsService
+    {
+        return this._breadcrumbsService;
+    }
+}